requirements.txt
Pillow
pystray
git+https://github.com/OperavonderVollmer/OperaPowerRelay.git